refactor(Nav): 내비게이션 바 파라미터 수정
