--- src/Main.js
@@ -1,22 +1,0 @@
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import App from "App";
-
-import { FolderPage, SharedPage, HomePage } from "./pages/index";
-
-function Main() {
-  return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<App />}>
-          <Route path="folder">
-            <Route index element={<FolderPage />} />
-            <Route path=":folderId" element={<FolderPage />} />
-          </Route>
-          <Route path="shared" element={<SharedPage />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
-  );
-}
-
-export default Main;
