+++ src/index.js
@@ -1,4 +1,4 @@
 import ReactDOM from "react-dom";
-import Main from "Main";
+import App from "App";
 
-ReactDOM.render(<Main />, document.getElementById("root"));
+ReactDOM.render(<App />, document.getElementById("root"));
